Split GUI toolbar playback tests
